Fixed the imeplemtations of sqlTableItem methods. Also added a connectionstring factory to SqlTableItem and SqlTableItemGroup

+++ Kode/DatabaseAPI/DatabaseAPI/DatabaseModel/Item.cs
@@ -2,11 +2,11 @@
 {
     public class Item
     {
-        public int ItemID { get; set; }
+        public long ItemID { get; set; }
         public string Name { get; set; }
-        public int ItemGroupID { get; set; }
+        public long ItemGroupID { get; set; }
 
-        public Item(int itemId, string itemName, int itemGroupId)
+        public Item(long itemId, string itemName, long itemGroupId)
         {
             ItemID = itemId;
             Name = itemName;
